% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeritos.R
\name{ampModelo}
\alias{ampModelo}
\title{Amplitude e Grau de Precisao}
\usage{
ampModelo(modelo, avaliando)
}
\arguments{
\item{modelo}{O modelo calculado por lm()}

\item{avaliando}{Dataframe com os dados do bem avaliando}
}
\value{
Valores de amplitude e Grau de Precisao
}
\description{
Calcula a amplitude e classifica quanto ao grau de precisao definido na NBR 14.653-2
}
\examples{
ampModelo(modelo, avaliando);
}
