% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeritos.R
\name{fpred}
\alias{fpred}
\title{Funcao para predizer valores de bens imoveis}
\usage{
fpred(model)
}
\arguments{
\item{model}{O modelo calculado por lm()}
}
\value{
Apresenta a funcao para precificacao de bens imoveis
}
\description{
Função para predicao de valores de bens imoveis via regressao linear
}
\examples{
fpred(modelo);
}
