% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeritos.R
\name{iConf_CA}
\alias{iConf_CA}
\title{Intervalos e Campo arbítrio}
\usage{
iConf_CA(modelo, avaliando)
}
\arguments{
\item{modelo}{O modelo calculado por lm()}

\item{avaliando}{Dataframe com os dados do bem avaliando}
}
\value{
Valores dos intervalos de confiança (superior e inferior) e o campo de arbítrio para o valor estimado do bem imóvel
}
\description{
Calcula os intervalos de confiança e o campo de arbítrio
}
\examples{
iConf_CA(modelo, avaliando);
}
