% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeritos.R
\name{micronum}
\alias{micronum}
\title{Funcao para testar micronumerosidade da amostra e indicar quantidade de elementos amostrais}
\usage{
micronum(data)
}
\arguments{
\item{data}{Dataframe do conjunto de dados}
}
\value{
Indica a quantidade ideal de elementos amostrais e classifica com o grau de fundamentacao correspondente
}
\description{
Função para calcular e classificar o grau de fundamentação quanto a micronumerosidade dos dados amostrais e indicar a quantidade ideal de dados
}
\examples{
micronum(data);
}
