% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeritos.R
\name{niveisRM}
\alias{niveisRM}
\title{Análise de Significancia das variaveis e do modelo}
\usage{
niveisRM(model, data)
}
\arguments{
\item{model}{O modelo calculado por lm()}

\item{data}{Dataframe dos dados}
}
\value{
Retorna a significancia das variaveis e do modelo e classifica referente ao Grau de Fundamentacao definido na NBR 14.653-2
}
\description{
Calcula a significancia estatistica das variaveis e do modelo
}
\examples{
niveisRM(model, data);
}
