% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeritos.R
\name{teste.chauvenet}
\alias{teste.chauvenet}
\title{Teste de Chauvenet}
\usage{
teste.chauvenet(dados)
}
\arguments{
\item{dados}{Vetor de dados com a variavel que sera analisada}
}
\value{
Retorna os valores do conjunto de dados que precisam ser removidos
}
\description{
Realiza o teste de Chauvenet para um vetor de dados.
}
\examples{
teste.chauvenet(dados$Valor)
}
