% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeritos.R
\name{ampModelo}
\alias{ampModelo}
\title{Amplitude e Grau de Precisão}
\usage{
ampModelo(model, ava)
}
\arguments{
\item{model}{O modelo calculado por lm()}

\item{data}{Dataframe com os dados do bem avaliando}
}
\value{
Valores de amplitude e Grau de Precisão
}
\description{
Calcula a amplitude e classifica quanto ao grau de precisão definido na NBR 14.653-2
}
\examples{
ampModelo(model, avaliando);
}
