% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeritos.R
\name{fpred}
\alias{fpred}
\title{Função para predizer valores de bens imóveis}
\usage{
fpred(model)
}
\arguments{
\item{model}{O modelo calculado por lm()}
}
\value{
Apresenta a função para precificação de bens imóveis
}
\description{
Função para predição de valores de bens imóveis via regressão linear
}
\examples{
fpred(modelo);
}
