% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeritos.R
\name{iConf_CA}
\alias{iConf_CA}
\title{Intervalos e Campo arbitrio}
\usage{
iConf_CA(modelo, avaliando)
}
\arguments{
\item{modelo}{O modelo calculado por lm()}

\item{avaliando}{Dataframe com os dados do bem avaliando}
}
\value{
Valores dos intervalos de confianca (superior e inferior) e o campo de arbitrio para o valor estimado do bem imovel
}
\description{
Calcula os intervalos de confianca e o campo de arbitrio
}
\examples{
iConf_CA(modelo, avaliando);
}
