% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeritos.R
\name{teste.dp}
\alias{teste.dp}
\title{Teste do Desvio Padrão}
\usage{
teste.dp(dados)
}
\arguments{
\item{dados}{Vetor de dados com a variável que será analisada}
}
\value{
Retorna os valores do conjunto de dados que precisam ser removidos e encontram-se acima ou abaixo de dois desvios padrões assim como o gráfico informando os pontos discrepantes.
}
\description{
Realiza o teste do desvio padrão para um vetor de dados.
}
\examples{
teste.dp(dados$Valor)
}
